requirements.txt
streamlit
python-dotenv
openai
heyoo
pygsheets
